requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-SocketIO==5.1.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
PyJWT==2.1.0
gunicorn==20.1.0
eventlet==0.33.0
python-iptables==1.0.0
Flask-Mail==0.9.1
